+++ NEWS.md
@@ -6,6 +6,8 @@
     -   url:
         <https://data.linz.govt.nz/layer/51768-nz-8m-digital-elevation-model-2012/data/>
 
+-   Add `volcano2` dataset
+
 # tidyterra 0.1.0
 
 -   Add DOI.